requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Babel==3.1.0
Flask-Caching==2.1.0
python-dotenv==1.0.0
psycopg2-binary==2.9.7
gunicorn==21.2.0
reportlab==4.0.4
Pillow==10.0.0
python-dateutil==2.8.2
waitress==2.1.2
email-validator==2.0.0
qrcode==7.4.2
openpyxl==3.1.2
Flask-SocketIO==5.3.6
python-socketio==5.11.2
python-engineio==4.10.1
langdetect==1.0.9
deep-translator==1.11.4
APScheduler==3.10.4
requests==2.31.0
pandas==2.1.1